pyproject.toml
[project]
name = "audio-diarizer"
version = "0.1.0"
description = "Audio diarization API server similar to AssemblyAI"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "redis>=5.0.0",
    "rq>=1.15.0",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "faster-whisper>=0.10.0",
    "pyannote.audio>=3.1.0",
    "librosa>=0.10.0",
    "ffmpeg-python>=0.2.0",
    "pydantic>=2.5.0",
    "pydub>=0.25.0",
    "numpy>=1.24.0",
    "aiofiles>=23.2.0",
    "sentencepiece==0.1.99",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
